[fidl][rust] Explicitly test send-side and receive side handle rights checks
Previous handle rights tests in rust only tested end-to-end handle
rights checking. This only demonstrated that the rights checking was
happening *somewhere*, but the requirement is to do checking on both the
send and receive side.

This CL adds separate testing of send side and receive side checking.
Because send-side is not implemented yet in rust, the tests are disabled
and currently do not pass when enabled. The technique for isolating send
side and receive side rights checks is to replace the ordinals during
transit over channels so that one protocol method is used on send side
and a different one on receive side - one with stricter rights checks
than the other.

Change-Id: Ibb9101236bf198e6a1b41b4aecffca91b361fa25
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/517802
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Fuchsia-Auto-Submit: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>