[fidl][llcpp] Fix handle leakage in client and server dispatch
Today, a FIDL incoming message's handle ownership is split into two
phases:

- When the message is in its encoded form (e.g. just received): the
handles are either owned by the |IncomingMessage| class, and freed
in bulk in an array via RAII, or manually freed.
- When the message is in its decoded form (e.g. after Decode()): the
handles are owned by a |fidl::DecodedMessage<FidlType>| sub-class, and
freed by traversing down the message - the handles are no longer in a
contiguous array.

It's easy to forget closing the handles, or lose track of ownership, if
the two phases are across one or more function calls. For example, the
server message dispatching logic decodes the incoming message by
looking up a list of |MethodEntry|s, and expects the method handler to
take over the ownership of decoded handles. However, that did not happen
in practice, so any handles not moved by the user would be leaked. A
similar issue exists on the client side when handling events.

This CL fixes the leaks using |fidl::DecodedMessage<T>| RAII types,
after pushing the decoding phase from the dispatcher runtime into the
generated code. The generated message handlers always take ownership
of handles in the incoming message.

TEST: fx test fidl_llcpp_dispatcher_handle_ownership_tests
Fixed: 74790
Change-Id: I661933678ef8cc819dc5f93ae1666a4fc52b139c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/514114
Commit-Queue: Yifei Teng <yifeit@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>