[bt][hfp] Send custom Battery Levels through test API.
Use the HFP test FIDL API to send battery level values to the HFP
component. This simulates the battery level changing on the Fuchsia
device. It will cause the HFP component to send a battery level status
indicator to the peer if the status indicator is activated.

The previous version of hfp::call_manager_init_and_peer_handling
incorrectly expected the the peer task to make a request for network
information. That's beyond the intended scope of the test which was
meant to test only that the hfp component was handling and routing
events in its main loop. It lead to failures when the order in which
the select! loop handled ready futures. Now the test is robust to
any ordering.

Test: SL4F test commands succeed.
Test: PTS receives test command.

Change-Id: I8d89e63828a1747774c5df9d8d9277df54798ebc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/515335
Fuchsia-Auto-Submit: Jeff Belgum <belgum@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Marie Janssen 💖 <jamuraa@google.com>