[vulkan_loader] Fix flake in LoaderUnittest.GoldfishDevice
goldfish_device.CloseAll() could be run before the pseudo-dir processed
the open request for the device. In that case, the device<->loader
channel wouldn't be closed and the test would hang waiting for the
loader device count to equal zero.

Removing the "dev" entry from the pseudodir will ensure that future
open requests for the device will fail, closing the channel and allowing
the loader to detect that the device went away.

Multiply: vulkan_loader_tests
Bug: 74558
Change-Id: If1c6d5482bdfdc5216438ad628c416e6a30c6b4b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/519482
Reviewed-by: John Rosasco <rosasco@google.com>
Commit-Queue: John Bauman <jbauman@google.com>