[cs] Add error propagation for entries() and exist().
Previously, when runtime exists in exec but is not being served, cs would crash without giving much context of why it fails. With this error propagation, it will give context on where it failed. Another CL will be out to gracefully exits with code 0 instead of reporting error in the case that part of a component doesn't exist.


Test: fx test cs-tests
      fx test cs-unittests

Fixed: 74655

Change-Id: I0eabb6be99f9101df204c0f98b0cf69d0009f935
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/515725
Reviewed-by: Xyan Bhatnagar <xbhatnag@google.com>
Reviewed-by: Fady Samuel <fsamuel@google.com>
Commit-Queue: Ingrid Feng <ingridlol@google.com>