[fxfs] Improve node lifecycle management
Removes a (very) coarse lock on the node cache, which was previously
held for the duration of disk loads on cache misses.

Implement NodeCache::insert_or_reserve() which returns a placeholder to
the first caller, and blocks all future callers until the placeholder is
filled (or the placeholder is dropped). This is used in
FxVolume::get_or_load_node to block all readers until the node is
loaded.

Change-Id: Iefc2131915c35013f17d53b9fb878397193fc666
Multiply: fxfs-tests: 100
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/516920
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>