[bt][hci] Fix TrySendNextQueuedPackets
Fix a bug in ACLDataChannel where we might not send all possible packets
in a call to TrySendNextQueuedPackets(), due to incorrectly testing the
head of the queue for technology type, instead of the current iterator.
this means that if one type (e.g. Acl or LE) reaches 0 availability, we
won't send any more packets even if there are available packets of the
other type that we _could send_.

Bug: fxbug.dev/74880
Test: fx test bt-host-hci-tests
HCI_ACLDataChannelTest.QueuedAclAndLePacketsAreSentUsingSeparateBufferCounts

Change-Id: Iea2ed80782355727dafabfebde1b83e026816999
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/503782
Reviewed-by: Marie Janssen 💖 <jamuraa@google.com>
Reviewed-by: Xo Wang <xow@google.com>
Commit-Queue: Xo Wang <xow@google.com>