[remote-control] regular pkg with one shell cmd
A small omission from an earlier cleanup change, the intent was to
have the package be "regular", where most binaries in the package are
not exposed as shell programs, only the runner. Having both
fuchsia_shell_package and the shell_command dependency resulted in a
conflict of intermediate outputs causing null build regressions for
certain build orders.

Change-Id: Ie49fc9a835fd36db3ceb6b614fd3a6b53c72321b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/519904
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: James Tucker <raggi@google.com>