[kernel][crashlog] Use FILE objects during crashlog recovery
When creating is filling out the VMO which will hold a recovered
crashlog, use a kernel FILE object to render to instead of using a
callback and a context pointer.

This change eliminates the need for a moderately sized stack buffer
when creating the crashlog header, and also will make it a bit easier
easier to start to incorporate other sources of data in the rendered
crashlog in future CLs.

Change-Id: I4d45c988f47aa01cea0e927fd47b10d91854cf4f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/506929
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>