[fidl][summarize] Harmonize fidl_api_summarize with RFC-076
In the course of reviewing what now is RFC-076, it was decided to add
the values of constants, enums, bits and struct elements with default
value to the API summary.

This change modifies the program code to produce matching summaries.

Tests are modified to match; tests are added to verify the new use cases
for value changes.

Bug: 72179
Change-Id: I062fb0f5b563e4f662b816cb7870588959640636
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/519981
Commit-Queue: Filip Filmar <fmil@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>