[virtualization][guest] Send errors of guest binary to stderr.
Currently, errors of the "guest" CLI binary are often sent to logging,
instead of the CLI. This may lead to user confusion, due to silent
failure of the command.

Modify guest to output to stderr instead of logs, given that the guest
binary is a CLI tool.

Change-Id: I46e1f657ef02a17ed970ae850a7a9ffefdc626d7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/519867
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: David Greenaway <dgreenaway@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>