[run-test-suite] Define interface for persisting scoped artifacts
Defines and leaves a space for an interface for persisting artifacts
scoped to a test run, suite, or test case to a directory.
Writing results "live" to stdout is not included as part of this
interface.

Up next: Add a preliminary implementation for writing to directories.

Change-Id: Iea86ff4f1c655580d1f4779d6e080f9110968aa4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/513992
Commit-Queue: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>