[volume_image]: Fix Off By One.
The problem was that data_slice slice(the first slice that is not backed
up by data and is not filled) was not falling into the right branch,
since the comparison ignored it (> data_slice_count) instead of >=.

Then on the next iteration, it will add the range including data_slice.
This was tricky since most of the time we have zero filled, and even
sometimes the order of the partitions made it so this wouldnt be
detected.

In the end, we can reproduce by having a mapping with no fillings, and 3
slices, with one backed by data.

Added an assert that would blow up without this fix.

Also, this tickled a bug in MtdWriterTest, since it now has empty slices
not incorrectly written as before, which has a different behavior on
nand and a file. A truncated file will assume that an unwritten offset
as this jump from the current end of file to further ahead is zero. On
the other hand, on a device, such as raw nand this is tread as unwritten
spaces. Because we are using a FTL, this ends up meaning that a range
is unmapped, which means that the contents are the same of an 'erased'
page, that is 0xFF.

Test: storage-volume-image-test
Change-Id: I8544626f668e7cf2e6dc8cc8e909e8bdab8e1375
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/520060
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Chris Suter <csuter@google.com>