[input] Fix comments regarding "machine uptime"
The `event_time` field corresponds to the monotonic time as determined
by the zx_clock_get_monotonic syscall.

This is not the same as machine uptime. It does not always reset on
reboot and does not adjust during sleep, and thus should not be used
as a reliable source of uptime.

https://fuchsia.dev/fuchsia-src/reference/syscalls/clock_get_monotonic
has more information on this.

Test: N/A
Change-Id: I9fea3fd015e3047559e7444eb9ace3d52d5dd665
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/519480
Reviewed-by: Mukesh Agrawal <quiche@google.com>
Reviewed-by: Alice Neels <neelsa@google.com>
API-Review: Alice Neels <neelsa@google.com>
Commit-Queue: Xyan Bhatnagar <xbhatnag@google.com>