[graphics] Support loading ICDs from packages.
The loader will connect to components to get information about their
ICDs, including a new manifest.json file. These ICD files can be used as
a fallback if the ICD shared library doesn't currently exist in
/system/lib. This will allow a soft transition from ICDs in the system
image to ICDs in a file.

The loader also implements an ICD runner, which is a simple way for a
driver package to serve its package contents. The interface with the ICD
runner is described in README.md.

Bug: 13078

Change-Id: I68439a153dec131992e478909288498fb024ce52
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/515640
Commit-Queue: John Bauman <jbauman@google.com>
Reviewed-by: Randall Bosetti <rlb@google.com>