[fidlc] Add element_memcpy_compatiblilty to vector coding table
element_memcpy_compatibility indicates if the data in the vector body
is possible to memcpy to the channel message, i.e. it doesn't require
any mutations. This new property element_memcpy_compatibility is
calculated off available type shape data

This will be used in the iovec encoder to detect when messages can
be directly pointed to by an iovec that is sent to the kernel. If the
encoder needs to mutate data, then performing these mutations on an
object that is not copied would break the invariant that the input
object isn't mutated (other than to move handles).

Bug: 66977
Change-Id: I66e1e1eadc797d8ba206b353d98d5c33f63154f1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/519380
Reviewed-by: Pascal Perez <pascallouis@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
API-Review: Pascal Perez <pascallouis@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>