[optee] Move OpteeClients to separate async loop
The optee driver was creating additional device instances for each
fuchsia.tee.Application and fuchsia.tee.DeviceInfo binding. This CL
creates a separate async::Loop with a single thread for servicing these
bindings. The existing deadline profile is applied to the new thread, as
it will be servicing trusted application requests. The OpteeDeviceInfo
class was eliminated in favor of serving the fuchsia.tee.DeviceInfo
protocol directly from the OpteeController class, as there is no server
state for that protocol.

Additionally, a test was added to ensure that any open TA sessions are
closed at the destruction of an OpteeClient, so that all OP-TEE
session resources are appropriately released.

Change-Id: Id4d1fffc2743ddbc96426a8cc99c1f6b2fbe9b2e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/513167
Commit-Queue: RJ Ascani <rjascani@google.com>
Reviewed-by: Dmitry Yatsushkevich <dmitryya@google.com>
Reviewed-by: Jonathan Bailey <jonathanbailey@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>