[blobfs] Fix tests in new pager.
Uses the new BlobfsTestSetup in more places.

Converts Fsck and BlobfsChecker to take a Blobfs pointer rather that
ownership of the Blobfs via a unique_ptr. This is because the successful
running of the Blobfs destructor depends on processing notifications
about paging event cleanup posted to the message loop. The caller needs
to manage the blobfs lifetime to make sure the message loop has time to
run before Blobfs teardown.

Change-Id: I4948518bab0c49df3afcc83786d090e396d53427
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/519406
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Martin Lindsay <mlindsay@google.com>