[fidlc] Change type used to store resolved handle rights.
In FIDL, handle rights and size bounds are similar: they are
both Constants that fall back to a default when unspecified
(same rights, and max size respectively).

This CL updates the compiler representation of handle rights
through the compilation process to be more consistent with that
of size bounds: we add a specific type alias on the ConstantValue
type used to represent handle rights, and use a pointer to this
to refer to a "resolved" handle rights value instead of storing
the top level Constant.

Besides adding consistency, by storing the ConstantValue rather
than the Constant we make it possible for the owned default value
to be `const static`, ensuring that there's a single copy
of the default for each of sizes and handle rights. Though this
is initially unnecessary (the lifetime of the TypeTemplates is
effectively static), this will not be the case when moving these
default values onto the respective Types as part of
fxbug.dev/74193

fx ninja -C out/default host_x64/fidl-compiler &&
./out/default/host_x64/fidl-compiler

Change-Id: I8da350d6938906d993b6e67d0f42de961631a591
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/517821
Reviewed-by: Pascal Perez <pascallouis@google.com>
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Commit-Queue: Felix Zhu <fcz@google.com>