[kernel][thread] make sure the Thread structure is aligned
Change Thread::CreateEtc to use memalign instead of malloc. Assert that
any pointers passed in are already properly aligned.

This lets some of the inner architectural structures be more strictly
aligned than usual and removes the need for some padding on x86. Also
increased the alignment of the fpu save state on arm64 to optimally
allow for dumping of the vector state.

Future enhancements: Remove most of the use of passed in Thread pointers
to CreateEtc and switch from directly driving malloc and use new and
AllocChecker. The latter requires some amount of restructuring of how
the constructor and init routines are implemented than a simple patch.

Tested: boot on both architectures, run full kernel unit tests, which
should exercise most of the exotic Thread creation spots in the kernel,
other than ThreadDispatcher.

Change-Id: I286f1fad7d725ee66a32162d790ae395102ce223
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/519545
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: John Grossman <johngro@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: Travis Geiselbrecht <travisg@google.com>