[image assembly] Add support for additional arbitrary HashDescriptors
Adds the ability to create arbitrary HashDescriptors, and a raw builder which
eases the creation of said HashDescriptors.

Adds a field for the minimum required version of avb that a hash descriptor
requires, and when creating a vbmeta image, sets the version based on the maximum
of all required minimum versions.

Bug: 73186

Change-Id: I67eae2e01d34eeae3cd334703c39a6cabe8d2a59
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/519985
Commit-Queue: Aaron Wood <aaronwood@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Reviewed-by: Allison Pearce <ampearce@google.com>