[kernel] Update JobDispatcher to report exited immediately on job kill
Currently it is possible for userspace to observe zx_info_job_t.exited =
false after a zx_task_kill call on a job if the job has not yet finished
terminating. The documented behavior of zx_info_job_t.exited is that
return_code is valid, which it is even when the job is in
State::KILLING. The current behavior is also inconsistent with
processes, which report exited = true immediately after a zx_task_kill.

This change will assist in debugging ZX_ERR_BAD_STATE errors returned by
fuchsia.process.Launcher that we theorize are from the job that a
process is being launched in terminating in the middle of it being
launched, which is an unavoidable race and not an indication of a deeper
problem.

Bug: 73781
Change-Id: I713f6ca6fbb03c2d94fec6dbbfb976f08e7e4173
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/519666
Fuchsia-Auto-Submit: Bryan Henry <bryanhenry@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Carlos Pizano <cpu@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>