[stdcompat]: Add support for is_sorted.
Constexpr is_sorted is a requirement for boot options.
Since we will do the runtime switch to std, we follow the same pattern
as cpp20::is_sorted.
As a side note, added test that verify that the is_sorted behavior
matches the std::is_sorted at runtime.
This verifies that the is_sorted treats < operator, as a <= comparison,
that is:
!comp(a[i+1], a[i]) where comp is the property being checked.

Also updated check to verify that the iterator category is a child of
the required type.

Test: stdcompat-test, stdcompat-unit-test(host)
Change-Id: Icea4311d0d3abb2cf5aae0037aba3acb5c2895e2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/519507
Reviewed-by: Corey Tabaka <eieio@google.com>
Fuchsia-Auto-Submit: Gianfranco Valentino <gevalentino@google.com>
API-Review: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>