[dart_tool] Use relative path in metadata
The main problem was that `snapshot` can be a fully qualified GN path,
for example something like `//out/default/...`. Somewhere downstream
this path can be written to a depfile, and it'll be treated as an
absolute path like `/out/default/...`, at least when it goes through
python, which happens to be the problem causing the linked bug.

Also a few other cleanups since I'm here. Note only the line rebasing
`snapshot` in metadata is fixing the bug.

Bug: 75437
Change-Id: I7806c806f4610176fa34b50849e390302b22373e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/520923
Commit-Queue: Jay Zhuang <jayzhuang@google.com>
Reviewed-by: Shai Barack <shayba@google.com>