[amlogic_decoder] Apply deadline profiles for H264
This CL applies deadline profiles to the various threads used for H264
multi instance decoding in the amlogic decoder driver. This will later
transition to requesting profiles based on the thread's "role" as
described in fxbug.dev/40858, so it introduces the concept of roles
with the expectation that it will be replaced with a different schema
for describing them as part of the aforementioned bug.

Change-Id: I61af5ee1c4f06980364208222771d5cfa37bcf83
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/515382
Commit-Queue: RJ Ascani <rjascani@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Reviewed-by: Dustin Green <dustingreen@google.com>