[fint] Fail if no-op check fails
https://fxrev.dev/508218 moved the ninja no-op check from recipes into
fint. In in doing so, it introduced a bug in the ninja no-op check
logic: `fint build` does a ninja dry run and parses the ninja stdio to
check for the "no work to do" string, but then proceeds to ignore the
result, emitting it to the build artifacts but not failing the
subprocess. In turn, the recipe only fails the build if the `fint build`
subprocess has a non-zero retcode.

No-op builds recently became broken as a result of this bug, so this fix
will not be safe to reland until the no-op build is fixed
(https://fxrev.dev/520923).

Bug: 75437
Change-Id: I321323aad444f8b0fdc355c19990b304ce517fe6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/520581
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Ina Huh <ihuh@google.com>
Reviewed-by: Shai Barack <shayba@google.com>