[fint] Emit `ninja explain` stdio in case of no-op failure
Recipes can then read these logs and present their contents to make it
easier for users to debug failed no-op checks. We used to do this when
the no-op check was implemented in recipes, but I removed it for
simplicity when rewriting the no-op check in fint.

Bug: 75437
Change-Id: I1fe46bf8db9b0b759889def08c41d04cead9481a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/521180
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Ina Huh <ihuh@google.com>