[fidl][rust] Use write_unaligned in fidl_struct_copy encode
This CL fixes potential undefined behavior in the encoding method for
`fidl_struct_copy!`, which was calling std::ptr::copy_nonoverlapping
a pointer that was non guaranteed to be properly aligned. This replaces
it with a write_unaligned call, which should compile down to the same
thing without any possibility of undefined behavior.

Test: fx test //src/lib/fidl/rust
Change-Id: I3d99d19f7a59095bf678bab77b0f2297389f7291
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/521137
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Commit-Queue: Mitchell Kember <mkember@google.com>