[debugger] Fix file name index lookup.
Clang now generates line table file names with index 0 and
non-normalized file names. Previous fixes for these issues missed the
index file which resulted in failures if a file name was only referenced
in this manner.

To test this, the checked-in symbol file has been updated. This resulted
in updates for some hard-coded constants used in some test expectations
which are now updated. This also revealed another deficiency in our
DWARF 5 implementation which is filed separately in issue 75488.

Some changes were made to the build ID index test to make the checked-in
symbol file easier to roll.

Bug: 75450

Change-Id: If44a1a739d01119f4a483d7e2ec3338eff6db52b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/520465
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>