[virtualization][arm64] Improve PL031 emulation.
The PL031 RTC is provided to guests to allow them to determine the wall
time. Currently, Linux is unable to use the PL031 because we fail to
handle reads/writes to the PL031's control register.

This CL updates the our simulated PL031 to support basic reads/writes to
the control register. We also update the update it to more gracefully
handle unknown reads/writes, logging a warning instead of returning an
error (which, in turn, causes the entire VM to crash).

Bug: 74914
Change-Id: Ie967ab5c0541764fc73734599a77a737a4c74504
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/520861
Commit-Queue: David Greenaway <dgreenaway@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>