[virtualization] Update mklinux.sh to gcc-10.
GCC version 9 is not longer packaged with Debian, making it harder for
users to obtain. Instead, update the script to use GCC 10 which is
actively supported.

We previously did not support to GCC 10, but patches to the Linux kernel
now allow us to successfully build:

  https://zircon-guest-review.googlesource.com/c/third_party/linux/+/2322
  https://zircon-guest-review.googlesource.com/c/third_party/linux/+/2323

Change-Id: Ic859c2e2fdc023980d06d087ac0af87c849e7e32
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/520864
Commit-Queue: David Greenaway <dgreenaway@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>