[virtualization][vmm] Move arch-specific VCPU code into arch/ dir.
Move the architecture-specific HandleMem, HandleInput, HandleOutput,
and HandleIO functions into the arch/ directory. Additionally, prepend
the name "Arch" to the functions to make it a little clearer that these
are per-architecture.

These functions are starting to get a little on the larger side, and the
amount of cross-arch code sharing is minimal. Having per-arch copies
allows us to simplify them, as in the next CL.

Change-Id: I51481d9cc2f1520db745e92e4ccc1b79c8d77c12
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/520865
Commit-Queue: David Greenaway <dgreenaway@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>