[build] Introduce renamed_binary() template.
This CL introduces a new template which will be used in the
few cases where an executable binary must be installed under
a different name inside a package.

The template ensures that the binary is installed to the
right location, _and_ that its runtime libraries are also
installed properly, something that doesn't work with a
simple resource() target.

Moreover, this provides the option of making the renamed
binary visible to the shell, but not the original binary
(which still must be installed). See documentation comments
for details and examples.

This will help in later CL in getting rid of misc. package()
instances (e.g. under //src/sys/pkg/bin/...).

Bug: 65068
Change-Id: Id13078a5ba4919f7d8849c443a17285ee55f598c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/519300
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: David Turner <digit@google.com>