[fidlcat] don't exit if no IR files
If fidlcat can't load the FIDL IR files, it now prints a warning and
allows the execution.

In that case, fidlcat is still able to decode all the syscalls but some
or all the FIDL messages may not be decoded and only dumped in binary.

Bug: 72663
Change-Id: Icb60db79d4bad341c6b29cdd21017a358b6e56b8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/521281
Commit-Queue: Vincent Belliard <vbelliard@google.com>
Reviewed-by: Francois Rousseau <frousseau@google.com>