[amlogic_decoder] Apply deadline profiles for VP9
This CL applies deadline profiles to the various threads used for VP9
decoding in the amlogic decoder driver. This will later transition to
requesting profiles based on the thread's "role" as described in
fxbug.dev/40858, so it introduces the concept of roles with the
expectation that it will be replaced with a different schema for
describing them as part of the aforementioned bug.

Change-Id: I3a4fbef26dc9389168910bf88e5fda378e94556a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/518242
Commit-Queue: RJ Ascani <rjascani@google.com>
Reviewed-by: John Bauman <jbauman@google.com>