[kernel] Panic quickly for unexpected page faults in kernel mode
The only kind of page fault that should ever happen in kernel mode
is for a data access during explicit user-copy operations.  For any
other case, a quick panic with full details of the fault site is
crucial to reducing the potential cascade harm of the bug and to
identifying the underlying root cause of the fault.

Bug: 66141
Change-Id: I4e4dce26f3e14733d7be7c07012aafca586f53c5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/459941
Testability-Review: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Venkatesh Srinivas <venkateshs@google.com>
Commit-Queue: Roland McGrath <mcgrathr@google.com>