[core] Route virtualization caps to console-launcher.
Currently, attempting to run the virtualization tool `guest` (which is
used to launch and query virtualized guests) fails when launched from
serial or an on-screen console because the
`fuchsia.virtualization.Manager` cap has not been routed to
`console-launcher`.

This CL adds the appropriate routing from `appmgr` to `console-launcher`
to avoid this problem.

Change-Id: Ibfe8b2a7e2206ba9b3f2d2006f0d2b924ac9d53a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/521322
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: David Greenaway <dgreenaway@google.com>