[zircon][kernel] Validate ZBI dcfg structs and make types portable.
This CL adds size and offset validation to the ZBI driver configuration
structs. As part of this, we add explicit padding to the structs, and
switch to uint8_t instead of bool to make sure that the size is always 1
byte on all platforms.

Bug: 74907
Change-Id: I3c2a85e5cb6bd9a12647f7f279f6f888663ff44f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/518925
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Simon Shields <simonshields@google.com>