[ffx doctor] match fx spawned daemons with config
As fx often now passes additional flags that are forwarded to daemon
invocations, those need to be accounted for in the pkill pattern.
We're likely approaching the end of the time where we should use
pkill/argument matching for this purpose, though a portable replacement
can be tricky.

Fixed: 75428
Change-Id: I725858807e97bed972c6e9ef6e3933cd064edf5f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/521283
Commit-Queue: Amit Uttamchandani <amituttam@google.com>
Reviewed-by: Jordon Wing <jwing@google.com>
Reviewed-by: Amit Uttamchandani <amituttam@google.com>