[fxfs] Implement rename
A fair bit of refactoring in unlink also happened, to try to reuse the
fiddly code paths in unlink as much as possible.  The general principle
that I aimed to follow was to construct the entire transaction in one
go, rather than in piecemeal calls (which was what I did first, but that
became error-prone as it was possible to construct logically
inconsistent transactions).  This necessarily leaked some details about
unlink semantics deeper into the object_store::directory code, such as
the existence of graveyard directories.

Tested: Unit, and `fx test rename-tests --test-filter=*/Fxfs`
Change-Id: I968c2d42461641b02ce0eb3e1a1107404e8aa073
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/520370
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Chris Suter <csuter@google.com>