[fake-clock] Loosen assertion on signals
Since the fake clock service immediately drops the server end of the
eventpair after signaling a stop point, it's possible for the
EVENTPAIR_CLOSED signal to also be present. This change filters out
EVENTPAIR_CLOSED when asserting on the returned signals.
The unit tests are not currently flaking, but I've updated them anyway
since they may break if something changes the order in which futures
are executed and dropped.

Fixed: 75629

Multiply: fake-clock-rust-v2

Change-Id: Ib9a855426551702158c2203e06ce54437bce1a58
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/522126
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>