[starnix] Organize fs crate
Move FidlFile and SyslogFile to a "fuchsia" subcrate to make room for
other file systems (e.g., procfs, tempfs).

Rename FidlFile to RemoteFile, which matches the terminology we use in
C++ for this concept.

Remove the Mutex around the SynchronousProxy because SynchronousProxy is
not thread-safe.

Rename FileDesc to FileObject because FileDesc was ambiguous between
FileDescriptor and FileDescription.

Rename FdNumber to FileDescriptor because that is the common industry
term for this type.

Rename FdHandle to FileHandle because its a handle to the actual file
rather than a handle to the file descriptor.

Finally, make the i32 inside FileDescriptor private. There is already a
raw() accessor for it.  We do not need to leak the structure of the type
itself.

Change-Id: Ia90c43c62c68cbb9a23e0368bae29c8d5ae7216e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/521906
Commit-Queue: Adam Barth <abarth@google.com>
Reviewed-by: Theodore Dubois <tbodt@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>