[scripts][clang] Add script for checking that vtables are in rodata
This is the script I used for checking if all vtables in a file, fuchsia
toolchain, llvm multilib, or fuchsia checkout are in the .rodata
section. The script works by naively parsing `readelf` output, finding
vtable symbols, and checking if they're in .rodata. It could be there's
a more production-friendly way of checking, but this is just what I
whipped up for local testing.

This is intended to be used by the clang canary builders as a final
check to ensure that vtables under the relative vtables ABI stay
readonly.

Bug: 45314
Change-Id: Id84246d31c7e033cb257f75efabaa803f1004156
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/513859
Commit-Queue: Leonard Chan <leonardchan@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>